1848 Fix bug where operating screws up if first skipped
If the first (or all) companies were skipped in an operating round, the
skip wasn't effective. The skip was checked in next_entity! only, which
is called only when going to the next company in the OR. Likewise, the
checking whether any companies were to run at all failed to take into
account whether those companies were skipped - meaning that the first
company in the OR would *kinda* run - it would run, but it wouldn't have
the things before it ran actually happen