18EU Pullman Trains
This train can be run to a single stop a different train is already visiting.

The constant utilized in 1825 allows ignoring the requirement to have a token on the route. This has been changed the a train parameter to allow the pulltrain train to utilize the same feature without being tied to 1825s mn requirement.