Implement variant DB1: additional upgrade paths for double dits
Unfortunately Dave Barry's kit DB1 is itself full of optional rules.
This implements the rule he "uses in practice", giving the 887/888
tiles to Units 1 and 3 but not 2 where they are primarily for the
more complex variant which he doesn't use in practice.

As with DB2 I have hence only implemented what I think is the most
popular version since 1825 has many options already.