[1858] Temporarily give majors start locations
Public companies do not have a fixed start location, this is determined
by the private company they are formed from or, if started after the
privates close, is any free city space.

Temporarily give them start locations so the game will run for the
moment, to let track/tokens/train operations to be implemented.