[1858] Check for legal tile placement, including gauge checks
The tile placement needs to be checked to make sure that it either adds
new track that is reachable by the company, or upgrades a city that the
company already has a route to. This has to be done twice, both for
routes using broad gauge track and for routes using narrow gauge track.