[1858] Make sure there is a legal route when placing a token
The checks in available_tokens meant that locations not legally linked
to an existing token were not shown as available. However, if you
clicked on a token spot that was only linked illegally (using both metre
and broad gauge track) you could still place a token there.

Overriding check_connected to check you have a route using either the
broad gauge graph or the narrow gauge graph fixes this.