[1858] Allow privates to lay unconnected plain track
This was being blocked as the track wasn't part of any route, so was
failing the check that the new track was usable by the company, even
though it was in their home hex. Fix this by just accepting anything
that a private company places in their home hexes, we don't have to
worry about anything being unreachable.