[1858] Corporations receive revenue for owned private companies
This isn't perfect yet as a public company that owns privates but no
trains (which will be usual in a company's first operating turn) will
have a routes step, but not have any trains to select routes for. The
user interation should be skipped in this case, though the entire step
cannot be as we need to log the revenue earned from the privates.