[1858] Fix actions available for private/minor companies
When run in hotseat mode, G1858::Company.company? was not inheriting the
expected value from Entity::Company and so some of the checks on
available actions were not being correctly evaluated.

Explicitly set the `company?` attribute to remove any differences in
behaviour between hotseat and server modes, and fix the actions.