[1858] Public companies can be started from private railways.
Everything mostly works, except when starting with a token in one of the
cities that is shared by two private railways. When that is done an
extra token slot is being created on the yellow tile, so you have one
slot with the public company token, and a second (empty) slot reserved
for the other private railway.

There are also a couple of UI glitches with this at the moment:
- Where two privates share a home city only one of them is shown on the
  map.
= There's a button for the user to swap a private railway for 50% of a
  treasury share in an unstarted company. This shouldn't be an option.