[1858] Don't set any coordinates for a public company
The corporation's coordinates were being set to avoid the player having
to choose its starting location if there was only one choice. But this
was causing problems in the graphing code as home_as_token is true --
any corporation that had its coordinates set to H11 would be able to
trace routes out of Madrid in any direction, not just from its token
there.

Fixes #20.