[1858] Check if a player can afford to start an auction
This speeds up stock rounds as a player's turn is now skipped if they
cannot take any actions in a stock round. Fixes #11.

This check had been disabled as it breaks the quick start variant as
nobody can act in the first stock round, but the game is not being
automatically advanced to the first operating round. This is issue #21.