[1858] Fix check for second tile discount
The new track gauges added were not being recorded for the first tile
laid in an operating turn, so the discount for narrow gauge track was
only being applied if this was on the second tile laid. This had been
working earlier, but the code adding the gauges to the round state had
been accidentally deleted when refactoring code (commit c7d5a018ff).

Fixes #17.