[1858] Skip track laying step if there are no available hexes
This makes a private railway skip its entire operating turn if it is not
able to lay any track. This should speed play up, as this happens a lot
after the first couple of operating rounds, and there's no point making
someone operate a private if all they can do is click 'pass'.

Fixes #8.