[1868WY] Golden/Northern Spike: 1-time bonuses for certain E-W runs
The spikes can each be only once, by completing particular E-W routes, and then
extending the route one additional town stop west to the spike location.

* Golden Spike: Omaha to Promontory Summit
* Northern Spike: any E to Billings to Independence Creek
* completing the spike gives a one-time revenue bonus, and then removes the
  Spike revenue location and upgrades the corresponding offboard tile(s) for
  Ogden or Billings (which has two separate locations on the map which both
  count as Billings)

In the unlikely event that a Railroad completes a Spike and has an additional
E-W route to Omaha/Bililngs, the Spike route is run first, and then the tile
are upgraded, and then the other route is run with the new revenue. Since we
have routes submitted all at once, the method `spike_new_end_tile_bonus`
computes the would-be bonus and reports to the user.

[#5011]