[1858] Stop extra city slots appearing in Sevilla/Córdoba/Zaragoza
Sevilla/Córdoba/Zaragoza each have two private railway companies that
have reservations on the city. Before the green tiles these cities only
have a single slot, but `find_reservation` was returning a slot index of
1 for the second company to be given the reservation. This was resulting
in the token going in slot 1 which shouldn't exist, but there were no
checks and so a second slot was created to put the token in.

Fixes issue #10.