[1858] Follow routes out of plain track in private home hexes
If a plain track tile was built in a private company's home hex, then
the graph computed did not follow this track, unless it was connected to
a town or city. See issue #7.

This commit creates a new node class to represent these plain track
paths. These are added to the node list in the graph and they are then
included in the walk list.

This might not be safe if this code was being used for evaluating train
routes, but this will only ever be considered for track lays by private
companies, or checking connections between private and major companies.