[1858] Add extra check for company/corporation connection
This is unlikely to come up in actual play, but no track is needed for a
connection between a private railway and a public company.

This can only happen in Sevilla or Córdoba, but it is possible to start
a public company in one of those cities without building and track
there. At that point there should then be a connection to the other
private company that has a home hex there.

Fixes #33.