[1858] Fix misleading log message
When converting a private railway company to a public company the log
message was something like `Player X exchanges Alar and Santander
Railway and Pt80 from the Treasury for a 40% share of N`, suggesting
that the money for half of the president's certificate was being paid
from the company treasury.

The game behaviour was correct (the money was being paid by the player),
the log message now matches this.

Fixes issue #9.