Allow revenue bonuses to be shown in the route selector
This is for where a corporation receives revenue from other sources as
well as its train route revenue. Allow these revenues to be shown in the
route selector so that the player has a clearer view of the
corporation's revenue sources, and the total revenue.

This has been implemented for 1858 where a corporation receives
additional revenue from any private companies it ownes, but will also be
useable by other titles, such as QLB's bonus revenue in 1873.