[1858] Reimplement private railways
Private railway companies are now represented by a pair of entitites: a
company and a minor.

The company is used to provide income (for players) or additional
revenue (for corporations) and reservations for token spots. It can be
part-exchanged for the president's certificate of a corporation.

The minor is used to lay track in operating rounds and controls the
track laid in the private railway's home hexes. It can be exchanged for
a share in a connected corporation.