[1858] Fix token placement and reservations in shared cities
There were still problems with handling tokens in cities that are
reserved by two companies. The previous method was working for the first
token placed on a yellow tile, but you could end up with the token and
the reservation sharing the same slot once the tile was upgraded to
green. This meant that another corporation could token the second slot
when it should have been reserved, or starting a corporation from the
second private would overwrite the existing token in the first slot.

Now the reservations are swapped if needed, to ensure that the correct
slot is reserved for the second company. This code is also moved from
find_reservation to exchange_token, so it is only called when the token
is being placed.

Fixes #43.