[1858] Don't exit PCR if there's a token to place
If the last private to operate in the private closure round was
exchanged for a treasury share, the player wasn't being given the option
to place a token.

Add a check for any pending tokens to ensure that the round won't finish
before the player gets a chance to place this token.

Fixes #41.