[1858] Don't pass after buying train if another can be scrapped
If a corporation bought a train that took them up to the train limit,
this was ending their train buying step. But if they have a scrappable
train (which might have been wounded by the train just bought) they
should be allowed to scrap that train and buy another.

Fixes #42.