[18Rhl] Fix CMV bug for RhE (fixes #8608)
Newly floated corporations should get a prise rise (bump) upwards
at the end of the stock round it is floated. Now RhE is floated as
part of the initial auction, so it did not get this prise rise.

Have now added a special case in the first SR to cover this case.
This also broke the reference game so it has been removed. Will
add a new reference game later.

Note! This do break existing 18Rhl games and wont be able to fix
them so I suggest archiving completed games, and pinning any running
game.