[1861] Corrected description of the game end timing
The game does not always finish after the next set of operating rounds.
If the first 8-train is purchased in the first round of an operating
round set then the game finishes at the end of the current set (which is
extended to three operating rounds).

This attempts to give a more accurate description of when the game ends.
It is far more verbose than the previous description.