[1861/1867] Minors/majors are nationalised in operating order
Both minor and major companies should be nationalised in operating
order. This can make a difference as if multiple corporations are being
nationalised it is possible for the national railway to run out of
tokens during the process, and it should take the tokens in the order
the corporations are nationalised.

The order of minor company nationalisation was based on the order of the
definitions in the game file. Majors were being nationalised in reverse
operating order.