[1858] Redo abilities for exchanging private for presidency
Change the implementation of the exchange ability for exchanging a
private railway company for the presidency of a public company.

There are now no changes to the core code, instead there are checks in
the 1858 to stop the company from being exchanged for a normal share of
an unfloated corporation.