[1858] Fix check for second tile metre gauge discount
The check for whether a second tile lay cost discount is given for metre
gauge track was incorrect. It was just checking whether the tile cost
was zero, but this cost includes any terrain costs, so the discount was
being off terrain cost for the first tile if it was a metre gauge tile.

Fixes issue #44.