[1858] Fix error if corporation closes due to EMR
The check to make sure a corporation buys a train after issuing shares
during emergency raising was not coping with a corporation closing if
enough shares are issued for it to hit the bottom of the stock market.
Check if this has happened and don't throw an error in this case.

Fixes the rest of issue #45.