[1858] Calculate private railway revenue in routes_revenue
The `routes` argument to `routes_revenue` is an empty array if no trains
are owned, which means it might not be possible to find the operating
corporation from it. For that reason the private railway revenue wasn't
being calculated in `routes_revenue` and a new method had been added to
Engine::Step::Dividend instead.

However it is possible to work out the currently operating corporation
from `@game.current_operator` so this change to the core code isn't
needed.  So let's do this calculation in `routes_revenue` instead.