[1861/1867] Don't move tokens from minor's charter to public company
When minor companies are merging the tokens on the map get moved across
to the new public company in the repeated calls to
`move_token(corporation, target)`.

This later call to `move_tokens_to_surviving` was normally not doing
anything as the tokens had already been moved. The exception is that if
any tokens have been removed from the map by `remove_duplicate_tokens`
because they were in the same city as a token of another of the merging
minor companies. These tokens are removed from the map and put back on
the minor's charter, and were then being moved across to the public
company as a zero cost token, messing up the cost of the public
company's tokens.

Removing this call to `move_tokens_to_surviving` fixes this.

This fixes one of the issues reported as tobymao/18xx#8589, where the
public company gets zero-cost tokens.