[1861/1867] Delete `round_end`, use version from base class
This method is only called from code that is determining the end of the
game, and it will never end after a merger round, so use the base class
version of this method which always returns `Round::Operating`.

This fixes tobymao/18xx#8660 which reported that 1861 would not finish
if the bank breaks in phase 2, as a merger round never happens.