[1858] Transfer reservations in the private closure round
This fixes #47.

In the private closure round, if one of the cities that is home to two
private railways (Sevilla, Córdoba or Zaragoza) still has a yellow tile,
then it was possible for the public company acquiring the private to be
unable to place a token there, if the other private hadn't quite closed
yet.

This was happening because the private being acquired was closing and
giving up its reservation before the tokening step. To avoid this
problem, transfer the reservation across to the public company.