[1858] Remove unneeded icons from city hexes
Private railway icons aren't needed in city hexes as their name is
visible in the reserved city slot. So get rid of the icons from these
hexes. The reservation name is only removed when a token is placed in
the city, which means that the private has been exchanged, so there's no
need to show its home hexes any more.

The exceptions are the three cities shared between two privates where
the token slot is still only showing the name of one of the companies.
The icons have been left in these hexes for the moment, at least until
something is done to fix the reservation display.