Catch all errors in `process_action`
In `validate.rb`'s `run_game()` function, the game engine is loaded outside of
the `begin`/`rescue` block, before calling `maybe_raise!` inside the block, so
that an error in the game state can be caught without crashing the whole
program.

The changes in the recent PRs #8735, #8756, and #8787 all changed/removed
certain game objects like trains and tiles. This caused the game engine to error
out when in `process_action` when converting the action to a hash (for example,
the 1852: Missouri change removed some trains from the roster, so actions using
the removed trains tried calling `@train.id` where `@train` was `nil`) before
`process_single_action` was called.

That made the `validate.rb` functions impossible to use to identify which games
to archive; the `NoMethodError` on the `nil` train crashed the whole thing.

An alternative change in `run_game()` would be to move the game load into the
`begin`/`rescue` block, but when the other kinds of errors are raised, the
`engine` ends up being `nil`, causing the code in the `rescue` block to just
raise further errors.