[1858] Token step in the operating round
In 1858 the cost of tokens are determined by the number of provincial
borders crossed from another token, measured as the crow flies.

The bulk of this commit is a hardcoded list of distances between cities.
It might be possible to calculate these distances by looking at the map
and hex borders, but this approach seemed far easier.

There are also base class methods overridden to ensure that there is a
legal route to the new token, using either narrow gauge or broad gauge
track.