[1858] Stock round: auctions, buying and selling shares
This is the start of the stock round implementation for 1858. Buying and
selling shares is fairly normal, the main difference being that you
can't directly par a corporation before phase 5. Before that point
corporations can only be started by exchanging a private railway
company, this is to be added in a later commit.

In a stock round you can also put up a private company for auction.
Everyone gets a chance to bid on this company, until only one person is
left. This is implemented using the Engine::Step::BuySellParSharesViaBid
code from 1861/1867.