[1858] Skip token step when exchanging if no available token slots
If a private railway company is being exchanged for a share, and it is
one that shares a city reservation with another private, then it is
possible that there will not be an available token slot at the point it
is exchanged (this can happen if the city is still yellow and the other
private company has already been exchanged and used to place a token in
the city). In this case skip the tokening step instead of forcing the
player to click the 'do not place token' button.

Fixes issue #51.