[1858] Can't exchange privates in the first stock round
Fixes issue 50. It wasn't possible to exchange a private for a share in
the first stock round, but the check for actions was thinking this was
allowed. If a player was left with enough cash after all the privates
had been sold to convert a private to a public company then they were
given another action instead of passing automatically.