[1858] Quick start and two player variants
The quick start variant skips the initial auction of private companies,
and instead randomly assigns players predefined batches of privates.

The two-player variant has seven privates unavailable for purchase
during the game, but still blocking track building and tokening. The
maximum percentage of a corporation that can be bought is increased from
60% to 70%.