[1858] Stock round: exchanging privates and placing tokens
Implementation of the stock round exchange and tokening steps.

Private railway companies in 1858 can be exchanged for shares in two
different ways:

1. A private with a city in one of its home hexes can be used to start a
   new corporation. The par price is taken from the private's face
   value, with the player paying for half of the president's
   certificate. The corporation starts with a token in the privates home
   city (if there are two cities in the private's home hexes then the
   player gets to choose which one to start in). The corporation becomes
   the owner of the private company.
2. A private can be exchanged for a treasury share from an existing
   corporation, if there is a valid route from one of the corporation's
   tokens to any of the privates home hexes. The corporation may place a
   token in a city in one of the private's home hexes. The corporation
   becomes the owner of the private company.

Any privates gained by corporations in this way will close at the end of
the corporation's dividend step in the next operating round. The
corporation will receive the private's face value as cash to its
treasury.

Later in the game, a corporation may be started by simply setting a par
price and purchasing the president's certificate. In this case the
corporation may start in any open token slot, and pays twice the city's
revenue to do this.

There's one place this commit touches common code. This is in
`Engine::Part::City.exchange_token`. There are three cities in 1858
which each have two private companies that have a reservation ability
for a token slot. However there is only a single slot in these cities
before they are upgraded to green. This change to `exchange_token`
catches the case where the reservation ability for the non-existent
second slot is being used, and switches the two reservations around.