Partially revert 4c65c1f, move changes to 1880 code
The changes introduced by commit 4c65c1f to
`lib/engine/step/home_token.rb` had changed game behaviour in 1867. The
home token for a minor needs to be placed before the minor has a
president, previously the player was the active entity during the home
token step, this had changed it to be the presidentless minor.

This removes this change and instead puts it in a G1880 subclass.

Fixes tobymao/18xx#8952.