[1861/1867] Test if final phase is entered after the bank has broken
Fixes tobymao/18xx#8862.

In both 1861 and 1867, if the bank breaks then the game will end at the
end of the current operating round. However, if the final phase is
reached (with the first 8T bought) after the bank has broken but before
the end of the same operating round, then that game ending rule takes
precedence, with the game ending after two or three more ORs (1861) or
at the end of the next set of ORs (1867).