Raise an error if the tile being laid (or its opposite) is already on the board
It's #9029 again, but the error is only raised if the tile's hex is the hex from
the game. Tiles need to be assigned a hex belonging to the TileSelector, so it's
not enough to check whether a tile has a hex, we need to check if that hex is
the game's hex.