[1858] Remove graph call from PrivateClosure.actions
The call to `exchange_corporations` in the actions method meant that the
graph had to be evaluated whilst a game's history is evaluated,
increasing the amount of work to be done.

This call has been removed from actions and instead `auto_actions` is
used to generate a pass if there is no exchange possible.

This change was requested in the discussion on PR tobymao/18xx#8967.