[1858] Remove graph call from Step::Track.actions
The call to `can_lay_tile?` in the actions method meant that the
graph had to be evaluated whilst a game's history is loaded,
increasing the amount of work to be done.

This call has been removed from actions and instead `auto_actions` is
used to generate a pass if it is not possible for a minor to lay any
track.