[1858] Corporations can grow up as a stock action
As a sell-type stock action, corporations (public companies) can convert
from a 5-share company to a 10-share company. And, at the start of phase
5, all companies that have not already grown up are automatically
converted.

To handle conversion as a stock action:
- Action::Convert is extended to include a corporation attribute.
- View::Game::BuySellShares has a `render_convert` method added, which
  adds a button for each corporation that is able to be converted.

There's also some 1858-specific code added to allow closed corporations
to be restarted, and to make sure they restart as 10-share companies if
we're after the start of phase 5.