[1867] don't allow a Major with no valid home to start; fix Toronto/Montreal starts
Fixes #8980

Prior to this fix, 1867 games could select a Major to start with no valid home,
and then be unable to place its home token. They were forced to undo the actions
instead of the game preventing the illegal move, e.g.,
http://18xx.games/game/115992?action=486

This PR changes the UI so that an option to start a Major is not presented, but
it does not change any action validation, so no previous games will break.

1861 already has its own override of `ipo_type`, so those games won't be affected.

From the designer in the comments on #8980:

> In 1867, minors may start in unconnected revenue centres in the same hex, but
> majors may not (if one such centre is occupied). Therefore, it should be
> impossible to start a major in Toronto since there is always a CN token in
> it. In principle, a major could start in Montreal (but that would be a bizarre
> situation).

A few 1867 games need pins due to Majors illegally starting in Toronto or
Montreal when another city on those hexes had a token:

`[96823,103457,113060,85612,91793,93176,116910,81411]`