[1858] Alternate view for minor corporations
The default view for a minor shows:
1. Its name.
2. The trains it owns.
3. The cash in its treasury.
4. Its position in the operating order.
5. The value of its last run.

Most of these are irrelevant for 1858, as the minor parts of the private
railway companies can't own trains, don't have a treasury, and don't run
routes.

This alternate view makes them look like private companies. This shows
more relevant information:
- Revenue.
- Home hexes.
- Whether it can be used to start a corporation (public company).
- Face value.
Making them look like companies also helps hide the internal split of
private railways into a company and a minor, this does not need to be
visible to the player.