[1858] Don't check graph in Step::BuySellParShares.actions
The call to `can_exchange_any?` in `Step::BuySellParShares.actions` was
leading to a call to `@game.corporation_private_connected?` which uses
the game graph. This will be slowing down loading the game history, so
this commit adds a simplified check which is just checking that shares
are available for the exchange.

When the game is being run interactively, `auto_actions` will check the
game graph to see if any exchanges are actually possible, and pass if
not.