[1822 family, core] add `lay_count` and `upgrade_count` to `tile_lay` settings
Fixes #9085, which only references the 1822 family, but the new ability
configuration is available for any game.

Will need to pin approximately 45 games which took advantage of the situation
described in #9085: a Major used the extra tile lay private--which is supposed
to grant one upgrade _or_ two new yellow lays--to lay a yellow and then do an
upgrade.