Call City.remove_reservation from Ability::Reservation.teardown
In Ability::Reservation.teardown the reservation was being removed by
calling City.reservations.delete. This could trigger a bug if there were
other reservations and also tokens in a city with multiple slots.

The bug appears if:
- The city has two slots, which are reserved by companies P1 and P2.
- A corporation has used P1's ability to place a token in slot 0.
- The city's reservations will be [P1, P2], so P2 has a reservation on
  the open slot.
- If P1 closes then teardown will call reservations.delete(P1), changing
  the reservations array to be [P2], so P2's reservation is now for slot
  0, which already has a token, and slot 1 is no longer reserved.
- By calling City.remove_reservation!(P1) instead, the reservations
  array will become [nil, P2], so P2's reservation is preserved.

A couple of 1846 tests are tweaked, as the reservations array returned
is [nil] instead of [] after this change.