Add hook to allow consent check to be made when buying shares
This is needed for 1858, which requires consent when a private company
is exchanged for a treasury share.

A new method is added to the Game class, `consenter_for_buy_shares`.
This should return the player who needs to consent to the action. The
default implementation always returns nil, which means that consent is
not needed.