[1858] Check consent when exchanging for a treasury share
When a private railway company is exchanged for a treasury share,
consent is needed from the president of the corporation if that is not
the same player as the one who owns the private company.

If the exchange is for a market share (only possible in the private
closure round) then consent is not needed.