Create a :future track type
Several games have pre-determined routes on the map where track will be
built later in the game:

- Harzbahn 1873 has concession routes for the railway companies.
- 1840 has the route for the Stadtbahn companies.
- 18SJ has three mainline routes.

On the physical game boards for these games these are indicated by lines
on the map. 18xx.games currently indicates these routes by having icons
in the hexes that form the route, but this is less clear as it does not
show the edges that the route enters and exits each hex.

This commit adds a :future track type for path tile parts. This gives a
dotted line on the game map, allowing the route to be drawn.

Any path tile parts with a track type of :future are not addded to the
tile's `paths` array, instead they are stored in a separated
`future_paths` array. This means that they will not be considered when
calculating routes, and are not checked or preserved when a tile is
upgraded.