Allow stubs to have different track gauges
The stub track gauge only makes a difference as to how it is rendered.
The code that enforces path placement with stubs does not pay any
notice of the track gauge.

This also adds a transparent border for :future track paths, so stubs
can be rendered if they have :future gauge.