[1858] Compact reservation lists on company closure
If a company closes that shares a reservation for a single slot city,
then compact that city's reservations array to make sure that the other
company's reservation is in the first slot. Without this it would be
possible that the reservation would be pointing to the non-existent
second slot, and so not shown on the map.