[1858] Structure choice actions instead of using label
The choice actions in the private closure round had only been specifying
a single value for the choice options, which was being used as both the
button label and the value stored in the 'choice' attribute of the
choose action. This value was then being parsed to extract the
information from it.

This instead passes a choice and a label, with the choice being a hash
with 'corporation' and 'from' attributes, containing the information
that identifies which share the private railway company is being
exchanged for. This should be less fragile than storing the button
label.