[1858] Don't check graph when checking if second tile can be played
There are checks in `can_lay_tile?` to see if there are no legal tile
placement actions. This is to allow a private railway company's entire
operating turn to be skipped if they cannot place any tiles.

To avoid slowing down game loading, these checks shouldn't be made when
loading a game as they involve computing the game graphs. This had been
handled correctly for an operator's first tile lay, but `can_lay_tile?`
was being called when checking after a lay_tile action, to see if
another tile can be placed. This was leading to calls to `Graph.compute`
when games were loaded, showing down the process.

This change skips these graphs checks when seeing if another tile can be
laid. This will mean that the track action won't be automatically
skipped if there are no legal tile lays, but the player has just laid
the first tile, so making them manually pass is no big deal.