[1858] Add explanation text for private exchange choices
The rules as to what shares can be swapped for a private railway company
are a little complicated in terms of what permissions need to be sought.
Add a fuller explanation of what needs to be requested from the
presidents of the corporations whose shares are being claimed in
exchange for the private.