[18Ardennes] Reimplement forts as hex assignments not hex tokens
The previous use of hex tokens for a corporation's fort token
assignments for a wasn't working with the two fort tokens in Metz (J18).
This is has been redone using Ft[12345] codes for the forts.

Also the use of hex tokens for the fort tokens on the map has been
removed, instead assignments are used for these as well.