[1861/1867] Fix share-buying order after merger
After multiple minors owned by different players are merged, if there is
a tie for directorship then this goes to the player whose minor was
first in operating order, not the one who started the merger process.

This is stated in section 12.3 of the second (All-Aboard Games) edition
rules. The third (Grand Trunk Games) edition rules are different
(section 9.2c) state that it is the owner of the active minor who gets
the first chance to buy more shares, but the designer has stated that
this is incorrect.

Fixes tobymao#9536.