Fix map rendering with TileSelector for Opal 1.6
Fixes #9605 for Opal 1.6 (progress for #9608); the issue is not present with
Opal 1.5.

The `selector` variable is set here with an `if` block, like this:

```
selector =
  if ...
    ...
  end
 ```

In Opal, this is compiled as `selector = (function() { ... })();`.

The problematic line is:

```
return h(:div) if select_tiles.empty?
```

In Opal 1.5, this `return` is incorrectly applied to the function wrapping the
`if` block for defining `selector`:

```
if ($truthy(select_tiles['$empty?']())) {
  return self.$h("div")
};
```

That bug is fixed in Opal 1.6, and the line is compiled so that the `render`
function is the one that returns, causing the whole map to be rendered as an
empty `<div>`, instead of only the `TileSelector` being rendered as an empty
`<div>`:

```
if ($truthy(select_tiles['$empty?']())) {
  $t_return.$throw(self.$h("div"))
};
```

So, with this change, the map and `TileSelector` still behave correctly in Opal
1.5, and will now work as intended in Opal 1.6.