[Opal 1.6] don't use return inside of a lambda
[opal/opal #2357](https://www.github.com/opal/opal/pull/2357) revamped how
closures are tracked. As described in #9673, this is better behavior and
unveiled some problems in our code, but Opal itself has a bug here (at least in
1.6.1), where a `return` inside of a lambda causes the enclosing function,
instead of the lambda itself, to return. Fortunately, in a lambda, `next`
behaves the same as `return`, so we can work around this issue by using `next`
in our lambdas.

Fixes #9603, #9604, #9606 for Opal 1.6--progress on #9608